+++ mingling/src/lib.rs
@@ -79,8 +79,6 @@ pub mod macros {
     pub use mingling_macros::completion;
     /// Used to create a dispatcher that routes to a `Chain`
     pub use mingling_macros::dispatcher;
-    /// Used to create a dispatcher that routes to a `Renderer`
-    pub use mingling_macros::dispatcher_render;
     /// Used to collect data and create a command-line context
     pub use mingling_macros::gen_program;
     /// Used to create a `Node` struct via a literal