+++ mingling_core/src/program/setup.rs
@@ -1,5 +1,3 @@
-use std::fmt::Display;
-
 use crate::{ProgramCollect, program::Program};
 
 mod basic;
@@ -13,7 +11,6 @@ pub use general_renderer::*;
 pub trait ProgramSetup<C, G>
 where
     C: ProgramCollect,
-    G: Display,
 {
     fn setup(&mut self, program: &mut Program<C, G>);
 }
@@ -21,7 +18,6 @@ where
 impl<C, G> Program<C, G>
 where
     C: ProgramCollect,
-    G: Display,
 {
     /// Load and execute init logic
     pub fn with_setup<S: ProgramSetup<C, G> + 'static>(&mut self, mut setup: S) -> S {